third_party/WebKit/LayoutTests/platform/win7/external/wpt/mediacapture-streams/MediaStreamTrack-getSettings.https-expected.txt

This is a testharness.js-based test.
FAIL A device can be opened twice and have the same device ID assert_equals: expected "b4450024-e949-437f-892c-d2d032a32350" but got "1102ebd1-3a8e-4e9a-bf7b-e565fcb7eeb9"
FAIL A device can be opened twice with different resolutions assert_equals: expected "a55a65d3-11c6-4499-a741-4864ff07c7ef" but got "7321437a-e101-4b26-b798-583b82e93fcd"
Harness: the test ran to completion.
